Add option to use custom Azure Storage endpoint
Currently storage-azure hard-codes the endpoint of the Azure Storage
account to `blob.core.windows.net` which means that the library can't be
used in regions such as Azure China or Azure Government, with Azure
Stack, or with the Storage Emulator for testing.

This change enables customizing the storage endpoint to remove this
limitation. With this change, we can for example use storage-azure to
talk to the Azurite storage emulator [1] like so:

```ts
const az = new AzureBlobStorage({
  container: process.env.CONTAINER,
  accountName: 'devstoreaccount1',
  accountKey: 'Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==',
  endpoint: 'http://localhost:10000'
})
await az.writeFile('azure-write/test.txt', 'using azure storage emulator')
```

[1] https://github.com/Azure/Azurite